Build adjustment.
I couldn't get it to run at first. I had
to change the CMake file to explicitly export
the 'cwrap' function for it to run.

I'm also getting two errors in the generated Javascript.
The touch-related functions fail to initialize and
one of the __findEventTarget calls fails.

I don't have a touch environment so I bypassed the touch functions.
For the __findEventTarget JS function, if the input is 0 I just
swap in '#canvas' and it works (or at least, fails in a way that doesn't
crash the application).

I am running in debug mode while I work on these things, and performance
is abysmal. I'm getting 20-30 fps with Firefox using 100% of a CPU core,
while native Craft gave me 60 fps with 16% of a CPU core.  I am hoping
that I can fix those two bugs and then see how release builds perform.